Moved Mike's classes into an executors package
